Support linking to headers in README
