+++ src/components/head-config.mustache
@@ -2,10 +2,10 @@
 <meta charset="UTF-8">
 <meta name="viewport" content="width=device-width">
 <link rel="icon" type="image/x-icon" href="{{&config.site}}/favicon.png" />
-<link rel="stylesheet" href="{{&config.site}}/fontawesome-6.2.0/css/all.min.css"  type="text/css"/>
+<link rel="stylesheet" href="{{&config.site}}/fontawesome-6.2.0/css/all.min.css" type="text/css"/>
 <script>
 	// Allow included JavaScipt files to determine the root directory
 	// of this site.
 	window.siteRoot = `{{&config.site}}`;
 </script>
-<script src="{{&config.site}}/bundle.js"></script>
+<script src="{{&config.site}}/bundle-main.js"></script>
