+++ src/pages/plugin/[pluginName].mustache
@@ -131,6 +131,9 @@
 	</div>
 </main>
 
+{{! Initializes screenshots and postprocesses the README }}
+<script src="{{&config.site}}/bundle-plugin-page.js"></script>
+
 {{> plugin-documentation-link}}
 
 </body>
