+++ src/runtime/components/components.scss
@@ -140,11 +140,3 @@ pre code {
 		margin-left: 15px;
 	}
 }
-
-.anchor {
-	position: relative;
-	top: -100px;
-	display: block;
-	height: 0;
-	overflow: hidden;
-}
