+++ src/runtime/index.ts
@@ -1,7 +1,6 @@
 import PluginDataManager from './PluginDataManager';
 import initializeSearch from './search/initializeSearch';
 import initializeDownloadPage from './initializeDownloadPage';
-import initializePluginPage from './plugin/initializePluginPage';
 import 'bootstrap';
 
 import './style.scss';
@@ -10,11 +9,7 @@ import './style.scss';
 const siteRoot: string = (window as any).siteRoot;
 
 window.addEventListener('DOMContentLoaded', async () => {
-	// Initialize components that don't depend on plugin data
 	const page = (window as any).pageId ?? 'default';
-	if (page === 'pluginDisplay') {
-		initializePluginPage();
-	}
 
 	const pluginDataManager = await PluginDataManager.fromURL(
 		`${siteRoot}/pluginData.json`,
