+++ src/runtime/plugin/initializeScreenshotCarousel.ts
@@ -35,11 +35,4 @@ const initializeScreenshotCarousel = (screenshotsCard: HTMLElement) => {
 	new Carousel('#' + screenshotsCarousel.id);
 };
 
-const initializePluginPage = () => {
-	// Update screenshot cards
-	const screenshotsCards = document.querySelectorAll('#screenshots-container');
-	for (const screenshotsCard of screenshotsCards) {
-		initializeScreenshotCarousel(screenshotsCard as HTMLElement);
-	}
-};
-export default initializePluginPage;
+export default initializeScreenshotCarousel;
